yt-dlp: Consider language when comparing itags
In YouTube response data we have a different itag per audio quality, but
it stays the same for different languages of the same quality. For this reason
we have to consider both itag and language when matching streams, otherwise
init and index ranges might be incorrectly extracted from wrong format info.