meson: Auto generate and compile gschemas
In upcoming Clapper 0.10 version, it will be possible for enhancer plugins
to have configurable properties. In order for config to be persistent,
a "gchema.xml" compiled with "glib-compile-schemas" is required.

Considering that schemas must have the same keys, types and values as enhancer
properties, we can just autogenerate them at compile time using a script/app
that loads compiled enhancers, generates schema for each and compiles them.

This project already had a dependency on "libclapper" which is required to
compile enhancers anyway.