yt-dlp: Support custom "ytdlp" URI scheme
Allows using "ytdlp" URI scheme to try to extract ANY site with yt-dlp.

We probably cannot test all 1800+ sites that yt-dlp supports, so lets
keep support for "https" with "X-Hosts" data for the ones that were
tested, thus will be picked automatically without URI changes.

The rest will be as a "try and see if it works" with "ytdlp" scheme.
Once some fequently used site is confirmed to work, PRs for adding it
to "X-Hosts" are welcome.