yt-dlp: Fetch and add to tags preview image
Use Gio file APIs to fetch best quality thumbnail. Afterwards, create
an image sample from downloaded data and add it to tags as preview image.

This change allows MPRIS enhancer to use it as media artwork, which is
the same as what Chrome browser shows in its own MPRIS implementation.